Feat(frontend): ErrorMessage, PatternError 컴포넌트 분리
