+++ frontend/src/pages/login/PatternError.tsx
@@ -1,0 +1,15 @@
+import { FieldError } from "react-hook-form";
+
+const PatternError = ({ field }: { field: FieldError | undefined }) => {
+  return (
+    <>
+      {field?.message && (
+        <small role="alert" className="text-red-500">
+          {field?.message.toString()}
+        </small>
+      )}
+    </>
+  );
+};
+
+export default PatternError;
